Fix flaky test with Jenkins heisenbug (#62)
* Fix flaky test with Jenkins heisenbug

The bug was not actually on Jenkins, I encountered it in like 50% of
attempts to build the project using Maven from CLI.

The problem was, that `testUploadTiming` method did set flag `_discardBatchesAfterPersistentFailures` of `EventUploader`, but this flag was not set back in the teardown method (as other flags are). The flag is significant in the `testPersistentErrorHandling` test where it could cause failures due to messages are not removed.

The randomness of the failures was caused by the random order of the test execution.

* Clean up in LogsTestBase

There was unused field, so it got removed. The same field is in the `MockServer` itself, so it was probably moved there and not removed.